// packages/mavlet-core/src/clients/mavlet-client/MavletClientOptions.ts
import { AnalyticsInterface, Storage } from '@mavrykdynamics/mavlet-types'

/**
 * @internalapi
 */
export interface MavletClientOptions {
  /**
   * Name of the application
   */
  name: string

  /**
   * A URL to the icon of the application
   */
  iconUrl?: string

  /**
   * A URL to the website of the application
   */
  appUrl?: string

  /**
   * The storage that will be used by the SDK
   */
  storage: Storage

  /**
   * The analytics instance that will be used by the SDK
   */
  analytics?: AnalyticsInterface
}